Conversion from line output to dict output
Line output was a temporary construct to have something simple to start
with.  Now, it's time for dictionary output, which I expect I can use
for yaml output by way of using the merge() to combine directives.

This handles the simple cases - next I need to get a plan for dashed
line style lists and arrays.