Coalesce (#5)
Identify (the first case of) directives that are dependent on each other and
merge them thru a process called 'coallese'. Dependent directives are
prefiltered into buckets based on their interdependent key, then subsequently
unified into a single directives that represents the merged state of all the
dependent directives on that key. This single directive can then be
intermingled in a pool of OneToOne directives for subsequent output.