Load schema file from package data
By adding the file to package_data, it shows in the binary package.
But not the source package.
MANIFEST.in handles including it in the source package.
Source package is needed for tox, which does the following:
https://github.com/tox-dev/tox/issues/844#issuecomment-396444052
So not including the MANIFEST.in means that tests fail due to the file
not being included.