Fix debconf selections v2 (#24)
* Simpler and better debconf selection handling

I had a really bad and wrong assumption - non-installer directives don't
start with 'd-i'.  This starts fixes for that.  In particular, the
package name should be taken from the first token, not split out from
the second token.

* Simpler logic for Unsupported

With improved knowledge of how non-d-i directives get sent, the logic
for Unsupported directives can be greatly simplified.

* Ver 0.0.6