Proposal:  handle multiline d-i directives
Multiline directive (ie: directives split over lines using a '\' on the end of each line)
cause the generator to fail with a stacktrace.

For example, this will fail when the generator is run on it:

```
tasksel tasksel/first multiselect server
d-i pkgsel/include string \
    vim \
    rsync \
    wget \
    curl
```

Changes here will combine lines of that 'd-i pkgsel/include', stripping surrounding space, and joining them together with a single space, and the generator then works.  pkgsel/include is still unsupported, so it still doesn't get converted, bt at least it doesn't break the generator

idx is also handled to use keep track fo the line the directive starts on, and pass that on the `directives.append()` line.